requirements.txt
Django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.12
gunicorn==25.3.0
python-decouple==3.8
requests==2.32.3
